Disable secret type API based on the config

+++ components/secret-mgt/org.wso2.carbon.identity.secret.mgt.core/src/main/java/org/wso2/carbon/identity/secret/mgt/core/SecretManagerImpl.java
@@ -27,6 +27,8 @@ import org.apache.commons.logging.LogFactory;
 import org.wso2.carbon.context.PrivilegedCarbonContext;
 import org.wso2.carbon.core.util.CryptoException;
 import org.wso2.carbon.core.util.CryptoUtil;
+import org.wso2.carbon.identity.base.IdentityConstants;
+import org.wso2.carbon.identity.core.util.IdentityUtil;
 import org.wso2.carbon.identity.secret.mgt.core.constant.SecretConstants;
 import org.wso2.carbon.identity.secret.mgt.core.dao.SecretDAO;
 import org.wso2.carbon.identity.secret.mgt.core.exception.SecretManagementClientException;
@@ -226,6 +228,12 @@ public class SecretManagerImpl implements SecretManager {
     @Override
     public SecretType addSecretType(SecretType secretType) throws SecretManagementException {
 
+        String enableSecretTypeEndpoint = getSecretTypeEndpointConfig();
+        if (!Boolean.parseBoolean(enableSecretTypeEndpoint)) {
+            throw new SecretManagementClientException(
+                    SecretConstants.ErrorMessages.ERROR_CODE_NOT_IMPLEMENTED_SECRET_TYPE_API.getCode(),
+                    SecretConstants.ErrorMessages.ERROR_CODE_NOT_IMPLEMENTED_SECRET_TYPE_API.getMessage());
+        }
         validateSecretTypeCreateRequest(secretType);
         String secretTypeID = generateUniqueID();
         secretType.setId(secretTypeID);
@@ -245,6 +253,12 @@ public class SecretManagerImpl implements SecretManager {
     @Override
     public SecretType replaceSecretType(SecretType secretType) throws SecretManagementException {
 
+        String enableSecretTypeEndpoint = getSecretTypeEndpointConfig();
+        if (!Boolean.parseBoolean(enableSecretTypeEndpoint)) {
+            throw new SecretManagementClientException(
+                    SecretConstants.ErrorMessages.ERROR_CODE_NOT_IMPLEMENTED_SECRET_TYPE_API.getCode(),
+                    SecretConstants.ErrorMessages.ERROR_CODE_NOT_IMPLEMENTED_SECRET_TYPE_API.getMessage());
+        }
         validateSecretTypeReplaceRequest(secretType);
         String secretTypeID;
         secretTypeID = retrieveOrGenerateSecretTypeId(secretType.getName());
@@ -264,6 +278,12 @@ public class SecretManagerImpl implements SecretManager {
     @Override
     public SecretType getSecretType(String secretTypeName) throws SecretManagementException {
 
+        String enableSecretTypeEndpoint = getSecretTypeEndpointConfig();
+        if (!Boolean.parseBoolean(enableSecretTypeEndpoint)) {
+            throw new SecretManagementClientException(
+                    SecretConstants.ErrorMessages.ERROR_CODE_NOT_IMPLEMENTED_SECRET_TYPE_API.getCode(),
+                    SecretConstants.ErrorMessages.ERROR_CODE_NOT_IMPLEMENTED_SECRET_TYPE_API.getMessage());
+        }
         validateSecretTypeRetrieveRequest(secretTypeName);
         SecretType secretType = getSecretDAO().getSecretTypeByName(secretTypeName);
         if (secretType == null || secretType.getId() == null) {
@@ -281,6 +301,12 @@ public class SecretManagerImpl implements SecretManager {
     @Override
     public void deleteSecretType(String secretTypeName) throws SecretManagementException {
 
+        String enableSecretTypeEndpoint = getSecretTypeEndpointConfig();
+        if (!Boolean.parseBoolean(enableSecretTypeEndpoint)) {
+            throw new SecretManagementClientException(
+                    SecretConstants.ErrorMessages.ERROR_CODE_NOT_IMPLEMENTED_SECRET_TYPE_API.getCode(),
+                    SecretConstants.ErrorMessages.ERROR_CODE_NOT_IMPLEMENTED_SECRET_TYPE_API.getMessage());
+        }
         validateSecretTypeDeleteRequest(secretTypeName);
         getSecretDAO().deleteSecretTypeByName(secretTypeName);
 
@@ -600,5 +626,11 @@ public class SecretManagerImpl implements SecretManager {
             }
             throw handleClientException(ERROR_CODE_SECRET_TYPE_DOES_NOT_EXISTS, secretType.getName());
         }
+    }
+
+    private static String getSecretTypeEndpointConfig() {
+
+        return IdentityUtil.getProperty(IdentityConstants.ServerConfig.
+                ENABLE_SECRET_TYPE_ENDPOINT);
     }
 }
